# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexgohost Billing & Client Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getMyOrders
      tags: [orders]
      summary: Get my orders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/orders:
    get:
      operationId: getAllOrders
      tags: [orders]
      summary: List all orders (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}/approve:
    post:
      operationId: approveOrder
      tags: [orders]
      summary: Approve order (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel order (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /invoices:
    get:
      operationId: getMyInvoices
      tags: [invoices]
      summary: Get my invoices
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
  /admin/invoices:
    get:
      operationId: getAllInvoices
      tags: [invoices]
      summary: List all invoices (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /admin/invoices/{id}/mark-paid:
    post:
      operationId: markInvoicePaid
      tags: [invoices]
      summary: Mark invoice as paid (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice marked paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/pay:
    post:
      operationId: payInvoice
      tags: [invoices]
      summary: Pay invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayInvoiceRequest"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
...
  /migrations:
    get:
      operationId: getMyMigrations
      tags: [migrations]
      summary: Get my migrations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My migrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Migration"
    post:
      operationId: requestMigration
      tags: [migrations]
      summary: Request website migration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MigrationRequest"
      responses:
        "201":
          description: Migration requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Migration"
  /admin/migrations:
    get:
      operationId: getAllMigrations
      tags: [migrations]
      summary: List all migrations (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All migrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Migration"
  /admin/migrations/{id}/status:
    put:
      operationId: updateMigrationStatus
      tags: [migrations]
      summary: Update migration status (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMigrationStatusRequest"
      responses:
        "200":
          description: Migration status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Migration"
  /payments/stripe/create-intent:
    post:
      operationId: createStripePaymentIntent
      tags: [payments]
      summary: Create Stripe payment intent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StripePaymentIntentRequest"
      responses:
        "200":
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripePaymentIntentResponse"
  /payments/transactions:
    get:
      operationId: getTransactions
      tags: [payments]
      summary: Get transaction log
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Transaction log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        type:
          type: string
          enum: [hosting, domain, upgrade, renewal]
        itemName:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, cancelled, completed]
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - clientId
        - type
        - itemName
        - amount
        - status
        - createdAt
    CreateOrderRequest:
      type: object
      properties:
        type:
          type: string
          enum: [hosting, domain, upgrade, renewal]
        itemId:
          type: string
        notes:
          type: string
      required:
        - type
        - itemId
    Invoice:
      type: object
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        amount:
          type: number
        tax:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [unpaid, paid, cancelled, overdue]
        dueDate:
          type: string
        paidDate:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        createdAt:
          type: string
      required:
        - id
        - invoiceNumber
        - clientId
        - amount
        - total
        - status
        - dueDate
        - createdAt
    InvoiceItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
      required:
        - description
        - quantity
        - unitPrice
        - total
    CreateInvoiceRequest:
      type: object
      properties:
        clientId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        dueDate:
          type: string
        tax:
          type: number
      required:
        - clientId
        - items
        - dueDate
    PayInvoiceRequest:
      type: object
      properties:
        method:
          type: string
          enum: [stripe, paypal, manual]
        paymentMethodId:
          type: string
      required:
        - method
    PaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        transactionId:
          type: string
        redirectUrl:
          type: string
        clientSecret:
          type: string
      required:
        - success
...
    Migration:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        domain:
          type: string
        oldHostingProvider:
          type: string
        oldCpanelHost:
          type: string
        oldCpanelUsername:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        progress:
          type: number
        notes:
          type: string
        requestedAt:
          type: string
        completedAt:
          type: string
      required:
        - id
        - clientId
        - domain
        - status
        - progress
        - requestedAt
    MigrationRequest:
      type: object
      properties:
        domain:
          type: string
        oldHostingProvider:
          type: string
        oldCpanelHost:
          type: string
        oldCpanelUsername:
          type: string
        oldCpanelPassword:
          type: string
        notes:
          type: string
      required:
        - domain
        - oldCpanelHost
        - oldCpanelUsername
        - oldCpanelPassword
    UpdateMigrationStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        progress:
          type: number
        notes:
          type: string
      required:
        - status
    Transaction:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        invoiceId:
          type: string
        amount:
          type: number
        method:
          type: string
          enum: [stripe, paypal, manual]
        status:
          type: string
          enum: [success, failed, pending, refunded]
        transactionRef:
          type: string
        createdAt:
          type: string
      required:
        - id
        - clientId
        - amount
        - method
        - status
        - createdAt
    StripePaymentIntentRequest:
      type: object
      properties:
        invoiceId:
          type: string
        amount:
          type: number
      required:
        - invoiceId
        - amount
    StripePaymentIntentResponse:
      type: object
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
      required:
        - clientSecret
        - paymentIntentId