# lib/api-zod/src/generated/types/paymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexgohost Billing & Client Management API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentResponse {
  success: boolean;
  transactionId?: string;
  redirectUrl?: string;
  clientSecret?: string;
}